+++ Sandboxie/core/dll/proc.c
@@ -973,7 +973,7 @@ _FX BOOL Proc_CreateProcessInternalW(
                     if(lpArguments)
                         wcscat(mybuf, lpArguments);
 
-                    if (! lpCurrentDirectory) {
+                    if (! lpCurrentDirectory) { // lpCurrentDirectory must not be NULL
                         lpCurrentDirectory = Dll_Alloc(sizeof(WCHAR) * 8192);
                         if (lpCurrentDirectory) {
                             ((WCHAR*)lpCurrentDirectory)[0] = L'\0';
@@ -992,7 +992,12 @@ _FX BOOL Proc_CreateProcessInternalW(
 
                     Dll_Free(mybuf);
 
-                    goto finish;
+                    //
+                    // when the service returns ERROR_NOT_SUPPORTED this means we should take the normal process creation route
+                    //
+
+                    if(err != ERROR_NOT_SUPPORTED)
+                        goto finish;
                 }
             }
         }
@@ -2298,7 +2303,7 @@ _FX BOOLEAN Proc_IsSoftwareUpdateW(const WCHAR *path)
     if (Dll_ImageType == DLL_IMAGE_MOZILLA_FIREFOX) {
 
         MatchExe = L"updater.exe";
-        static WCHAR* Dirs[] = { L"\\mozilla firefox\\updates\\" , L"\\mozilla\\updates\\", L""};
+        static WCHAR* Dirs[] = { L"\\mozilla firefox\\updates\\" , L"\\mozilla\\updates\\", L"\\mozilla firefox\\", L""};
         MatchDirs = Dirs;
         SoftName = L"Mozilla Firefox";