+++ SandboxiePlus/SandMan/SandMan.cpp
@@ -83,6 +83,25 @@ CSandMan* theGUI = NULL;
 
 extern QString g_PendingMessage;
 
+#include <QStyledItemDelegate>
+class CTrayBoxesItemDelegate : public QStyledItemDelegate
+{
+	void paint(QPainter *painter, const QStyleOptionViewItem &option, const QModelIndex &index) const
+	{
+		QStyleOptionViewItem opt(option);
+		if ((opt.state & QStyle::State_MouseOver) != 0)
+			opt.state |= QStyle::State_Selected;
+		else if ((opt.state & QStyle::State_HasFocus) != 0 && m_Hold)
+			opt.state |= QStyle::State_Selected;
+		opt.state &= ~QStyle::State_HasFocus;
+		QStyledItemDelegate::paint(painter, opt, index);
+	}
+public:
+	static bool m_Hold;
+};
+
+bool CTrayBoxesItemDelegate::m_Hold = false;
+
 CSandMan::CSandMan(QWidget *parent)
 	: QMainWindow(parent)
 {
@@ -220,6 +239,8 @@ CSandMan::CSandMan(QWidget *parent)
 	m_pTrayBoxes->setHeaderHidden(true);
 	m_pTrayBoxes->setSelectionMode(QAbstractItemView::NoSelection);
 	//m_pTrayBoxes->setSelectionMode(QAbstractItemView::ExtendedSelection);
+	//m_pTrayBoxes->setStyleSheet("QTreeView::item:hover{background-color:#FFFF00;}");
+	m_pTrayBoxes->setItemDelegate(new CTrayBoxesItemDelegate());
 
 	pLayout->insertSpacing(0, 1);// 32);
 
@@ -314,16 +335,11 @@ CSandMan::CSandMan(QWidget *parent)
 	OnStatusChanged();
 	if (CSbieUtils::IsRunning(CSbieUtils::eAll) || theConf->GetBool("Options/StartIfStopped", true))
 	{
-		SB_STATUS Status = ConnectSbie();
-		CheckResults(QList<SB_STATUS>() << Status);
+		SB_RESULT(void*) Status = ConnectSbie();
+		HandleMaintenance(Status);
 	}
 
 	//qApp->setWindowIcon(GetIcon("IconEmptyDC", false));
-
-	if (!g_PendingMessage.isEmpty()) {
-		OnMessage(g_PendingMessage);
-		PostQuitMessage(0);
-	}
 }
 
 CSandMan::~CSandMan()
@@ -347,7 +363,9 @@ CSandMan::~CSandMan()
 
 void CSandMan::LoadState()
 {
+    setWindowState(Qt::WindowNoState);
 	restoreGeometry(theConf->GetBlob("MainWindow/Window_Geometry"));
+	restoreState(theConf->GetBlob("MainWindow/Window_State"));
 	//m_pBoxTree->restoreState(theConf->GetBlob("MainWindow/BoxTree_Columns"));
 	m_pMessageLog->GetView()->header()->restoreState(theConf->GetBlob("MainWindow/LogList_Columns"));
 	m_pLogSplitter->restoreState(theConf->GetBlob("MainWindow/Log_Splitter"));
@@ -358,6 +376,7 @@ void CSandMan::LoadState()
 void CSandMan::StoreState()
 {
 	theConf->SetBlob("MainWindow/Window_Geometry", saveGeometry());
+	theConf->SetBlob("MainWindow/Window_State", saveState());
 	//theConf->SetBlob("MainWindow/BoxTree_Columns", m_pBoxTree->saveState());
 	theConf->SetBlob("MainWindow/LogList_Columns", m_pMessageLog->GetView()->header()->saveState());
 	theConf->SetBlob("MainWindow/Log_Splitter", m_pLogSplitter->saveState());
@@ -444,6 +463,7 @@ void CSandMan::CreateMenus()
 	m_pMenuOptions = menuBar()->addMenu(tr("&Options"));
 		m_pMenuSettings = m_pMenuOptions->addAction(CSandMan::GetIcon("Settings"), tr("Global Settings"), this, SLOT(OnSettings()));
 		m_pMenuResetMsgs = m_pMenuOptions->addAction(tr("Reset all hidden messages"), this, SLOT(OnResetMsgs()));
+		m_pMenuResetGUI = m_pMenuOptions->addAction(tr("Reset all GUI elements"), this, SLOT(OnResetGUI()));
 		m_pMenuOptions->addSeparator();
 		m_pEditIni = m_pMenuOptions->addAction(CSandMan::GetIcon("EditIni"), tr("Edit ini file"), this, SLOT(OnEditIni()));
 		m_pReloadIni = m_pMenuOptions->addAction(CSandMan::GetIcon("ReloadIni"), tr("Reload ini file"), this, SLOT(OnReloadIni()));
@@ -570,8 +590,10 @@ void CSandMan::closeEvent(QCloseEvent *e)
 				theConf->SetValue("Options/PortableStop", PortableStop);
 		}
 
-		if(PortableStop == 1)
-			StopSbie(true);
+		if (PortableStop == 1) {
+			SB_RESULT(void*) Status = StopSbie(true);
+			// don't care for Status we quit here anyways
+		}
 	}
 
 	QApplication::quit();
@@ -669,7 +691,7 @@ void CSandMan::OnMessage(const QString& MsgData)
 	{
 		QString Op = Message.mid(3);
 
-		SB_STATUS Status;
+		SB_RESULT(void*) Status;
 		if (Op == "Connect")
 			Status = ConnectSbie();
 		else if (Op == "Disconnect")
@@ -680,36 +702,8 @@ void CSandMan::OnMessage(const QString& MsgData)
 			Status = theAPI->TerminateAll();
 		else
 			Status = SB_ERR(SB_Message, QVariantList () << (tr("Unknown operation '%1' requested via command line").arg(Op)));
-		CheckResults(QList<SB_STATUS>() << Status);
-	}
-	else if (Message.left(6) == "Status")
-	{
-		QString Status = Message.mid(7);
-		if (Status != "OK")
-		{
-			if(m_bStopPending)
-				QMessageBox::warning(NULL, tr("Sandboxie-Plus - Error"), tr("Failed to stop all Sandboxie components"));
-			else if(m_bConnectPending)
-				QMessageBox::warning(NULL, tr("Sandboxie-Plus - Error"), tr("Failed to start required Sandboxie components"));
 
-			OnLogMessage(tr("Maintenance operation %1").arg(Status));
-			CheckResults(QList<SB_STATUS>() << SB_ERR(SB_Message, QVariantList() << Status));
-		}
-		else
-		{
-			OnLogMessage(tr("Maintenance operation Successful"));
-			if (m_bConnectPending) {
-
-				QTimer::singleShot(1000, [this]() {
-					SB_STATUS Status = this->ConnectSbieImpl();
-					CheckResults(QList<SB_STATUS>() << Status);
-				});
-			}
-		}
-		m_pProgressDialog->hide();
-		//statusBar()->showMessage(tr("Maintenance operation completed"), 3000);
-		m_bConnectPending = false;
-		m_bStopPending = false;
+		HandleMaintenance(Status);
 	}
 }
 
@@ -720,11 +714,11 @@ void CSandMan::dragEnterEvent(QDragEnterEvent* e)
 	}
 }
 
-void CSandMan::RunSandboxed(const QStringList& Commands, const QString& BoxName, const QString& WrkDir)
+bool CSandMan::RunSandboxed(const QStringList& Commands, const QString& BoxName, const QString& WrkDir)
 {
 	CSelectBoxWindow* pSelectBoxWindow = new CSelectBoxWindow(Commands, BoxName, WrkDir);
 	//pSelectBoxWindow->show();
-	SafeExec(pSelectBoxWindow);
+	return SafeExec(pSelectBoxWindow) == 1;
 }
 
 void CSandMan::dropEvent(QDropEvent* e)
@@ -1390,11 +1384,15 @@ void CSandMan::OnNotAuthorized(bool bLoginRequired, bool& bRetry)
 
 void CSandMan::OnBoxMenu(const QPoint & point)
 {
-	QTreeWidgetItem* pItem = m_pTrayBoxes->currentItem();
+	QPoint pos = ((QWidget*)m_pTrayBoxes->parent())->mapFromParent(point);
+	QTreeWidgetItem* pItem = m_pTrayBoxes->itemAt(pos);
 	if (!pItem)
 		return;
+	m_pTrayBoxes->setCurrentItem(pItem);
 
+	CTrayBoxesItemDelegate::m_Hold = true;
 	m_pBoxView->PopUpMenu(pItem->data(0, Qt::UserRole).toString());
+	CTrayBoxesItemDelegate::m_Hold = false;
 
 	//m_pBoxMenu->popup(QCursor::pos());	
 }
@@ -1452,10 +1450,9 @@ void CSandMan::OnDisableForce2()
 	theAPI->DisableForceProcess(Status);
 }
 
-SB_STATUS CSandMan::ConnectSbie()
+SB_RESULT(void*) CSandMan::ConnectSbie()
 {
-	SB_STATUS Status;
-	bool bJustStarted = false;
+	SB_RESULT(void*) Status;
 	if (!CSbieUtils::IsRunning(CSbieUtils::eAll)) 
 	{
 		if (!CSbieUtils::IsInstalled(CSbieUtils::eAll))
@@ -1475,24 +1472,16 @@ SB_STATUS CSandMan::ConnectSbie()
 				return SB_OK;
 		}
 
-		bJustStarted = true;
 		Status = CSbieUtils::Start(CSbieUtils::eAll);
 	}
 
 	if (Status.GetStatus() == OP_ASYNC) {
 		m_bConnectPending = true;
-		return SB_OK;
+		return Status;
 	}
 	if (Status.IsError())
 		return Status;
-	if (bJustStarted) {
-		QTimer::singleShot(1000, [this]() {
-			SB_STATUS Status = this->ConnectSbieImpl();
-			CheckResults(QList<SB_STATUS>() << Status);
-		});
-		return SB_OK;
-	}
-	
+
 	return ConnectSbieImpl();
 }
 
@@ -1505,6 +1494,11 @@ SB_STATUS CSandMan::ConnectSbieImpl()
 		return SB_OK;
 	}
 
+	if (!g_PendingMessage.isEmpty()) {
+		OnMessage(g_PendingMessage);
+		PostQuitMessage(0);
+	}
+
 	return Status;
 }
 
@@ -1513,9 +1507,9 @@ SB_STATUS CSandMan::DisconnectSbie()
 	return theAPI->Disconnect();
 }
 
-SB_STATUS CSandMan::StopSbie(bool andRemove)
+SB_RESULT(void*) CSandMan::StopSbie(bool andRemove)
 {
-	SB_STATUS Status;
+	SB_RESULT(void*) Status;
 
 	if (theAPI->IsConnected()) {
 		Status = theAPI->TerminateAll();
@@ -1535,7 +1529,7 @@ SB_STATUS CSandMan::StopSbie(bool andRemove)
 
 void CSandMan::OnMaintenance()
 {
-	SB_STATUS Status;
+	SB_RESULT(void*) Status;
 	if (sender() == m_pConnect)
 		Status = ConnectSbie();
 	else if (sender() == m_pDisconnect)
@@ -1566,11 +1560,56 @@ void CSandMan::OnMaintenance()
 	else if (sender() == m_pUninstallAll)
 		Status = StopSbie(true);
 
+	HandleMaintenance(Status);
+}
 
+void CSandMan::HandleMaintenance(SB_RESULT(void*) Status)
+{
 	if (Status.GetStatus() == OP_ASYNC) {
+
+		HANDLE hProcess = Status.GetValue();
+		QWinEventNotifier* processFinishedNotifier = new QWinEventNotifier(hProcess);
+		processFinishedNotifier->setEnabled(true);
+		connect(processFinishedNotifier, &QWinEventNotifier::activated, this, [processFinishedNotifier, this, hProcess]() {
+			processFinishedNotifier->setEnabled(false);
+			processFinishedNotifier->deleteLater();
+			
+			DWORD dwStatus = 0;
+			GetExitCodeProcess(hProcess, & dwStatus);
+
+			if (dwStatus != 0)
+			{
+				if(m_bStopPending)
+					QMessageBox::warning(NULL, tr("Sandboxie-Plus - Error"), tr("Failed to stop all Sandboxie components"));
+				else if(m_bConnectPending)
+					QMessageBox::warning(NULL, tr("Sandboxie-Plus - Error"), tr("Failed to start required Sandboxie components"));
+
+				OnLogMessage(tr("Maintenance operation failed (%1)").arg((quint32)dwStatus));
+				CheckResults(QList<SB_STATUS>() << SB_ERR(dwStatus));
+			}
+			else
+			{
+				OnLogMessage(tr("Maintenance operation Successful"));
+				if (m_bConnectPending) {
+
+					QTimer::singleShot(1000, [this]() {
+						SB_STATUS Status = this->ConnectSbieImpl();
+						CheckResults(QList<SB_STATUS>() << Status);
+					});
+				}
+			}
+			m_pProgressDialog->hide();
+			//statusBar()->showMessage(tr("Maintenance operation completed"), 3000);
+			m_bConnectPending = false;
+			m_bStopPending = false;
+
+			CloseHandle(hProcess);
+		});
+
 		//statusBar()->showMessage(tr("Executing maintenance operation, please wait..."));
 		m_pProgressDialog->OnStatusMessage(tr("Executing maintenance operation, please wait..."));
 		SafeShow(m_pProgressDialog);
+
 		return;
 	}
 
@@ -1586,12 +1625,14 @@ void CSandMan::OnViewMode(QAction* pAction)
 
 void CSandMan::OnAlwaysTop()
 {
+	StoreState();
 	bool bAlwaysOnTop = m_pWndTopMost->isChecked();
 	theConf->SetValue("Options/AlwaysOnTop", bAlwaysOnTop);
 	this->setWindowFlag(Qt::WindowStaysOnTopHint, bAlwaysOnTop);
 	this->show(); // why is this needed?
 	m_pPopUpWindow->setWindowFlag(Qt::WindowStaysOnTopHint, bAlwaysOnTop);
 	m_pProgressDialog->setWindowFlag(Qt::WindowStaysOnTopHint, bAlwaysOnTop);
+	LoadState();
 }
 
 void CSandMan::SetViewMode(bool bAdvanced)
@@ -1705,6 +1746,46 @@ void CSandMan::OnResetMsgs()
 
 	theAPI->GetUserSettings()->UpdateTextList("SbieCtrl_HideMessage", QStringList(), true);
 	m_pPopUpWindow->ReloadHiddenMessages();
+}
+
+void CSandMan::OnResetGUI()
+{
+	hide();
+
+	theConf->DelValue("ErrorWindow/Window_Geometry");
+	theConf->DelValue("MainWindow/Window_Geometry");
+	theConf->DelValue("MainWindow/Window_State");
+	theConf->DelValue("MainWindow/BoxTree_Columns");
+	theConf->DelValue("MainWindow/LogList_Columns");
+	theConf->DelValue("MainWindow/Log_Splitter");
+	theConf->DelValue("MainWindow/Panel_Splitter");
+	theConf->DelValue("MainWindow/BoxTree_Columns");
+	theConf->DelValue("MainWindow/TraceLog_Columns");
+	theConf->DelValue("FileBrowserWindow/Window_Geometry");
+	theConf->DelValue("FileBrowserWindow/FileTree_Columns");
+	theConf->DelValue("NewBoxWindow/Window_Geometry");
+	theConf->DelValue("OptionsWindow/Window_Geometry");
+	theConf->DelValue("OptionsWindow/Run_Columns");
+	theConf->DelValue("OptionsWindow/AutoRun_Columns");
+	theConf->DelValue("OptionsWindow/Groups_Columns");
+	theConf->DelValue("OptionsWindow/Forced_Columns");
+	theConf->DelValue("OptionsWindow/Stop_Columns");
+	theConf->DelValue("OptionsWindow/Start_Columns");
+	theConf->DelValue("OptionsWindow/INet_Columns");
+	theConf->DelValue("OptionsWindow/NetFw_Columns");
+	theConf->DelValue("OptionsWindow/Access_Columns");
+	theConf->DelValue("OptionsWindow/Recovery_Columns");
+	theConf->DelValue("OptionsWindow/Templates_Columns");
+	theConf->DelValue("PopUpWindow/Window_Geometry");
+	theConf->DelValue("RecoveryWindow/Window_Geometry");
+	theConf->DelValue("RecoveryWindow/TreeView_Columns");
+	theConf->DelValue("SelectBoxWindow/Window_Geometry");
+	theConf->DelValue("SettingsWindow/Window_Geometry");
+	theConf->DelValue("SnapshotsWindow/Window_Geometry");
+
+	LoadState();
+
+	show();
 }
 
 void CSandMan::OnEditIni()
@@ -1922,7 +2003,11 @@ void CSandMan::OnSysTray(QSystemTrayIcon::ActivationReason Reason)
 		{
 			QMap<QString, CSandBoxPtr> Boxes = theAPI->GetAllBoxes();
 
+			bool bActiveOnly = theConf->GetBool("Options/TrayActiveOnly", false);
+
 			bool bAdded = false;
+			if (m_pTrayBoxes->topLevelItemCount() == 0)
+				bAdded = true; // triger size refresh
 
 			QMap<QString, QTreeWidgetItem*> OldBoxes;
 			for(int i = 0; i < m_pTrayBoxes->topLevelItemCount(); ++i) 
@@ -1938,6 +2023,9 @@ void CSandMan::OnSysTray(QSystemTrayIcon::ActivationReason Reason)
 					continue;
 
 				CSandBoxPlus* pBoxEx = qobject_cast<CSandBoxPlus*>(pBox.data());
+
+				if (bActiveOnly && pBoxEx->GetActiveProcessCount() == 0)
+					continue;
 
 				QTreeWidgetItem* pItem = OldBoxes.take(pBox->GetName());
 				if(!pItem)
@@ -1981,6 +2069,8 @@ void CSandMan::OnSysTray(QSystemTrayIcon::ActivationReason Reason)
 
 				m_pTrayMenu->removeAction(m_pTrayList);
 				m_pTrayMenu->insertAction(m_pTraySeparator, m_pTrayList);
+
+				m_pTrayBoxes->setFocus();
 			}
 
 			m_pTrayMenu->popup(QCursor::pos());